Fully initialize Scroll pHAT HD when creating I2CDisplay
We now fully initialize the Scroll pHAT HD by:
 * Setting all LEDs on the controller that are hooked up to physical
   LEDs to "on".
 * Setting all LEDs to not blink.
 * Setting the PWM for all LEDs to a 0% duty cycle.

We do this for all three settings, as they have an undefined state
when the controller is first powered on.

Unfortunately, smbus_write_block_data reliably corrupts the data it
sends to the display even when sending a single byte, so we need to
use smbus_write_byte_data in all cases.